Simplify, accelerate the video provider, now `VideoServiceProcessor`
